refactor(apps/api): rename albums module to library-albums
Reorganize the albums feature module to align with the library
subfeature structure, consistent with the library-artists module.
This change moves all album-related commands, queries, DTOs, and
controllers under the library-albums namespace.
