+++ apps/api/src/features/features.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AlbumsModule } from './albums/albums.module';
 import { AuthModule } from './auth/auth.module';
+import { AlbumsModule } from './library-albums/library-albums.module';
 import { LibraryArtistsModule } from './library-artists/library-artists.module';
 import { LibraryModule } from './library/library.module';
