+++ apps/api/src/features/library-albums/commands/handlers/delete-album.handler.ts
