+++ apps/api/src/features/library-albums/commands/handlers/update-album.handler.ts
