+++ apps/api/src/features/library-albums/commands/handlers/upload-album-cover.handler.spec.ts
