+++ apps/api/src/features/library-albums/commands/impl/delete-album.command.ts
