+++ apps/api/src/features/library-albums/commands/impl/update-album.command.ts
