+++ apps/api/src/features/library-albums/commands/impl/upload-album-cover.command.ts
