+++ apps/api/src/features/library-albums/dto/create-album.request.dto.ts
