+++ apps/api/src/features/library-albums/dto/create-album.response.dto.ts
