+++ apps/api/src/features/library-albums/dto/get-album.response.dto.ts
