+++ apps/api/src/features/library-albums/dto/update-album.response.dto.ts
