+++ apps/api/src/features/library-albums/dto/upload-album-cover.request.dto.ts
