+++ apps/api/src/features/library-albums/dto/upload-album-cover.response.dto.ts
