+++ apps/api/src/features/library-albums/queries/handlers/get-album.handler.ts
