+++ apps/api/src/features/library-albums/queries/impl/get-album.query.ts
